apps/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
boto3>=1.34.72
aioboto3>=12.3.0
langchain>=0.1.4
langchain-aws>=0.1.0
openai>=1.10.0
anthropic>=0.8.1
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
redis==5.0.1
websockets==12.0

# IBM Cloud SDK for Triage Module
ibmcloudant>=0.8.0
ibm-watson>=8.0.0
ibm-cloud-sdk-core>=3.18.0

# Email and templating
jinja2>=3.1.2
